+++ src/components/todo/PageTitle.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PageTitle = ({ text }: { text: string }) => {
+  return <h1 className="text-4xl">{text}</h1>;
+};
+
+export default PageTitle;
